fix: hooks and nested disclosure
