+++ src/App.tsx
@@ -1,7 +1,6 @@
 import './App.module.css';
 import NavBar from './components/NavBar';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import { Home } from './components/Pages/Home';
 import { Layout } from 'antd';
 import { ContentWrapper } from './components/ContentWrapper';
 import { GithubOutlined } from '@ant-design/icons';
